+++ packages/runtime/datastore/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @fluidframework/datastore
 
+## 2.0.0-internal.5.0.0
+
+### Major Changes
+
+-   The `@fluidframework/garbage-collector` package was deprecated in version 2.0.0-internal.4.1.0. [8b242fdc79](https://github.com/microsoft/FluidFramework/commits/8b242fdc796714cf1da9ad3f90d02efb122af0c2)
+    It has now been removed with the following functions, interfaces, and types in it.
+
+    -   `cloneGCData`
+    -   `concatGarbageCollectionData`
+    -   `concatGarbageCollectionStates`
+    -   `GCDataBuilder`
+    -   `getGCDataFromSnapshot`
+    -   `IGCResult`
+    -   `removeRouteFromAllNodes`
+    -   `runGarbageCollection`
+    -   `trimLeadingAndTrailingSlashes`
+    -   `trimLeadingSlashes`
+    -   `trimTrailingSlashes`
+    -   `unpackChildNodesGCDetails`
+    -   `unpackChildNodesUsedRoutes`
+
+## 2.0.0-internal.4.4.0
+
+Dependency updates only.
+
 ## 2.0.0-internal.4.1.0
 
 ### Minor Changes